The result of fgetc is an int, not a char
Without this change, get_pid_environ_val would go into an infinite loop
when asked to find a variable that doesn't exist on a platform where
char is unsigned (e.g. ARM): fgetc would return -1 (EOF), which would be
stored as 255 in temp[i], which then wouldn't be equal to -1 when
testing.